[printing] Move CheckForCancel() to printing::mojom::PrintManagerHost
On the current code, even though |preview_ui_| is checked before
calling CheckForCancel(), the renderer could not get the reply as
the print WebUI dialog is already closed on the browser. If then,
the renderer gets blocked.

Since CheckForCancel() is called multiple times while handling the
print preview request and |preview_ui_| could be disconnected once
the print WebUI is closed, this change moves CheckForCancel() to
mojom::PrintManagerHost as PrintViewManager which implements it
is alive while WebContents is available.

Bug: 1166492, 1167170, 1008939
Change-Id: Iffc5d5354be3778ec7f33cff11051858432b5bf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2632023
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: Julie Kim <jkim@igalia.com>
Cr-Commit-Position: refs/heads/master@{#845636}