Avoid IPC spam and reentrancy when sending initial frame properties
When a site-isolated iframe moves to a new process, we need to ensure
that the browser sends it an initial set of properties that are
generated by the parent frame, even if those properties have not
changed. Previously, we redundantly re-sent the properties every time
frame rects for the iframe changed. With this patch, we send the
initial properties only once, after the first time VisualProperties
is sent to the frame.

This is a follow-up to:

https://chromium-review.googlesource.com/c/chromium/src/+/2628746

Bug: 1135714
Change-Id: I72e2f30d12525ab3f8cfb97a178b5185a9f6ca13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2640893
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845685}