[Autofill] ImportFromSection returns true if profile is importable.
Currently, the return value is conditions on the actual import of
an address. However, the import can only fail if the incognito mode is
enabled, which is already checked beforehand. Therefore, the return
value for an importable profile is set to be always true if the profile
is importable to support a consistent behavior with future asynchronous
import flows.

Bug: 1135178
Change-Id: I7f062151706919439c82a6bd150ec8be7f219f1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2632963
Commit-Queue: Matthias Körber <koerber@google.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845736}