TableView: Select-all works for 1+ items
This adds the ability in TableView to add a range of items comprised of
only 1 item. This requirement manifests itself when using a "select-all"
capability. Prior to this CL, only ranges of 2+ items would be allowed.

Furthermore, this fixes a bug when all items are selected while there is
no 'active' row in a table. This can occur when the selection UI is
outside of the TableView.

Change-Id: I950fca34e8c83b650d14e0d6da3e7ad9d0b84f33
Fixed: 1169260
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2641320
Commit-Queue: Olivier Yiptong <oyiptong@chromium.org>
Auto-Submit: Olivier Yiptong <oyiptong@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845741}