capture_mode: Fix video dimensions when display bounds changes
When recording a window or a partial region, the display can be
rotated, or its device scale factor can change, leading to the
DIPs display bounds changing. If we don't propagate this change
to the recording service, the video capturer will try to fit (i.e.
center and letter-box) the new frame dimensions into the original
requested video resolution. This is ok when recording fullscreen,
but leads to completely broken cropping when recording a region
(See bug for video demo). It can also lead to a bad quality
window recording.

This CL fixes this issue by propagating display bounds changes to
the service to reconfigure the video capturer and encoder if
necessary (Demo of fix is on the bug).

BUG=1168434
TEST=Manually, added a new test

Change-Id: I9db5d29bddc217042b007b422b1084c86d8be959
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2641082
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845749}