Pipe the TWA quality violation message to the DevTools issue panel
This goes through the RenderFrameHost to get to the Impl under content,
similar to the HeavyAdIssue. Then build up the message and send to
devtools protocol.

This CL also adds passing the package name and signature to jni when
violation type is DIGITAL_ASSET_LINK, so that they can be use for
displaying the issue.

Design doc
https://docs.google.com/document/d/1Dsw4bYttTODZJ7QbRlS4-iN7nBFnJJ1-JGmUY6gdMiE/
Issue string
https://docs.google.com/document/d/1nlYeWlwCOva7M9EoT6YKQZ4s-Ug8aXbQIp4Oc1cL8WE/

Bug: 1147479
Change-Id: I57b5fb9b415889efc998cf88223ef0a87b77ade8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2568749
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Commit-Queue: Ella Ge <eirage@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845779}