Don't send phishing model to non-existent frames
OnRenderFrameCreated will send to frames when they are created in the
renderer process - avoid sending the model to frames that do not exist.
Previously this was silently dropped, and in this case the code handles
things by sending the message later.

R=drubery@chromium.org

Bug: 1169066, 1158869
Change-Id: I7daffcce74c568f1c09f18288ab3e7cde3bc6b86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2642880
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Auto-Submit: danakj <danakj@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845786}