[CrOS Cellular] Add CellularInhibitor class
This class updates the "Inhibited" property of the Cellular device.
When some SIM-related operations are performed, properties of the
Cellular device can change to a temporary value and then change back.
To prevent churn in these properties, Shill provides the "Inhibited"
property to inhibit any scans.

This class is intended to be used when performing such actions to ensure
that these transient states never occur. Because this operation is
expected to be performed often, we encapsulate it within this class.

Note that this CL does not add any clients for the new class; this will
occur as part of follow-up CLs.

Bug: 1093185
Change-Id: I5776c5c4a47db3645951d639b2917d8c5769c45f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2640314
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Azeem Arshad <azeemarshad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845823}