Increase the priority of the chromium_presubmit builder.
The CQ for infra/config changes often only run the chromium_presubmit
builder and when there are outages that affect the builders that share
machines with chromium_presubmit, it means that changes to attempt to
address the outage may experience long pending times and/or require
skipping the presubmit. Increasing the priority will allow us to keep
the safety provided by the presubmit while still being able to land
changes quickly to address outages.

Bug: 1168201
Change-Id: I828ffcf74a162a6ebdaba0075e6d66ffb31ab2c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2643407
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845827}