Add WebappsClient::AddShortcut
This installs a homescreen shortcut for a webpage. It seems at least
some of the code for doing this, i.e. AddShortcutWithSkBitmap(), is
componentizable and could run in WebLayer. This may be desirable
specifically for the case where Chrome is not the default browser
(creating a shortcut in the default browser). However for now, that
decision and work is punted.

Bug: 1164069
Change-Id: I8fbb789dc833a09358a821aba828d7a82ee9fd86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2640935
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845837}