Clear the cache in AutomationManagerAura when disabled.
On the Tast tests which use the Automation features, I'm seeing
sometimes the automation node tree gets empty. During the debugging,
I've noticed that sometimes the automation tree fails to build
the tree updates (happens OnSerializeFailure).

This can be resolved by clearing the cache when it is disabled
(which happens regularly on tests).

Bug: 1165292
Test: on kevin64, run informational Tast tests
AX-Relnotes: n/a
Change-Id: I835e06ca2118ae6c36d6fb8b877fad074bd41fcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2627715
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845868}