Experimental CanvasFormattedText API
This change adds support for an experimental CanvasFormattedText API.
This API aims to make it easy for web developers to get correct and
fast multiline formatted text rendering functionality in Canvas,
while preserving the low-level flexibility for positioning text
that is already present in the existing Canvas API

More details and use cases for the API can be found here
https://github.com/MicrosoftEdge/MSEdgeExplainers/blob/main/Canvas/FormattedText.md

With this change an initial subset of the proposal is implemented.
fillFormattedText is introduced on the canvas rendering context that
allows for rendering multiline text in canvas.

Not yet implemented:

-The font and color attributes on the canvas text.
-Support for writing modes & text direction.
-Support for the advanced Advanced usage in the explainer.
-StrokeText
-Ability to mutate runs once they are appended to CanvasFormattedText.
-Tests will be in a follow up change.

Perf Characteristics:

Comparing with hypothetical javascript multiline implementation,
that is a call to measureText + fillText to draw a line.

CanvasFormattedText is faster to re-render the same text on both a
laptop / desktop class hardware.

When it comes to first render, depending on the number of lines
the text breaks into performance varies.

On the hardware tested it is 10% (laptop  i7-7660U) and
16% (desktop  i9-9940X) faster when there are 3 lines. When there is
just a single line and no line breaking involved CanvasFormatted text
is 70% to 50% slower than a call to measuretext + filltext.

Sample Code:
<SCRIPT>
    function DrawMultiline()
    {
        var context = document.getElementById("target")
            .getContext("2d");
        context.font = "30px Arial";
        var text = new CanvasFormattedText();
        text.appendRun({ text: " Lorem Ipsum " });
        text.appendRun({ text: " is sample text" });
        context.fillFormattedText(text, 100, 0, 300);
    }
</SCRIPT>
<BODY>
    <canvas width=500 height=300 id="target"></canvas>
    <button onclick="DrawMultiline()">Draw</button>
</BODY>

Change-Id: I31df1a8c538dc799ec1a328e4952d01e78bbc344
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2570327
Commit-Queue: Sushanth Rajasankar <Sushraja@microsoft.com>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845879}