LayoutShiftTracker: Reset post-prepaint state even if we early out.
This patch ensures that we reset some state that we accumulate and
only reset if we actually report a layout shift.

The problem is that the accumulated values can later cause us _not_
to early out from ObjectShifted and report a layout shift where we
should not have reported one.

By resetting the state after every prepaint, we ensure that we reset
the accumulated state.

R=skobes@chromium.org

Change-Id: I2e764d3d847669cbdf06219da183cda02df57684
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2639862
Commit-Queue: vmpstr <vmpstr@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845904}