[Autofill] Added language detection.
This CL registers a TranslationDriver::LanguageDetectionObserver
and re-runs Autofill's heuristics when the language has been detected.

Forms are extracted from the DOM and sent to Autofill before the
page language has been determined. Therefore, without the listener,
Autofill will be informed about the page language only after the
renderer proactively sends updated forms to the browser, which
happens upon certain user actions and certain dynamic changes of the
DOM (e.g., when the user focuses a field). By then, it may very well
be too late to re-run Autofill's heuristics.

This CL fixes this issue: Autofill is now notified about the page
language once it is determined by Translate. Autofill then re-runs
the heuristics for the specific language.

Note that this is about language determination, not about actual
translation. Autofill does not need notifications about translations.
Once the site is translated and a re-parse happens, Autofill will
use the translated language to run the heuristics, but this behaviour
is not affected by this CL.

This CL has a side-effect on repeated-field type rationalization, which
previously was only run when server predictions were received. This
appears to have been a bug in one way or another; it will be addressed
later. For now, this CL adjusts the test expectations (*.out files) to
take into account repeated-field type rationalization.

This CL builds on crrev.com/c/2575080.

Bug: 1150895, 1168239
Change-Id: Ie018b9281305895f74541245076c0958828e3de4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2561838
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Megan Jablonski <megjablon@chromium.org>
Reviewed-by: Matthias Körber <koerber@google.com>
Reviewed-by: Michael Bai <michaelbai@chromium.org>
Reviewed-by: John Wu <jzw@chromium.org>
Reviewed-by: Josh Simmons <jds@google.com>
Cr-Commit-Position: refs/heads/master@{#845908}