[Chromecast] Prepare BrokerService as a replacement for Service Manager
This allows us to eventually unify both internal and external service
interconnection under a single framework (BrokerService). This patch
creates some temporary hooks which make it easier for us to migrate from
Service Manager to BrokerService. As part of the transition, the
lifetime of the Cast BrokerService is no long scoped by the Service
Manager; it will instead by created eagerly in the Cast main process and
expose a service request handler so that clients can still use the
Service Manager to connect to it.

Merge-With: eureka-internal/513086

Bug: internal b/169098077
Test: CQ

Change-Id: I20960c2b57d2b2481c3fcac44b20ca7e3a658786
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2622577
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Kenneth MacKay <kmackay@chromium.org>
Commit-Queue: Sean Topping <seantopping@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845910}