Refactor ChromeDataExchangeDelegate::GetFilenames for reuse
Refactor only change.

Refactor GetFilenames to return internal FileInfo struct which
includes FilePath and FileSystemURL so that it can be reused by
CreateClipboardFilenamesPickle in crrev.com/c/2638967.

By keeping the FileSystemURL in addition to the FilePath, it saves
duplicating the code to crack the URL from the path.

Bug: b/162602338
Change-Id: I91356840303bc6f15e11fcf43cad75c06e29fe7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2643736
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Jason Lin <lxj@google.com>
Cr-Commit-Position: refs/heads/master@{#845944}