Prerender: Remove dependencies on RenderFrameHost from PrerenderHost(Registry)
This CL is a preparation for implementing the Omnibox trigger. This
removes dependencies on the initiator RenderFrameHost from
PrerenderHost(Registry) so that the Omnibox trigger that doesn't have a
concept of the initiator frame will be able to trigger prerendering.

Before this CL, PrerenderProcessor passed the identifier of the
initiator frame up to PrerenderHost in order to retrieve BrowserContext.
After this CL, StoragePartitionImpl passes BrowserContext when
instantiating PrerenderHostRegistry, so the registry and hosts no longer
need the initiator frame.

Bug: 1166085
Change-Id: I0da9a95427c10ef65be7f2539311d50c002d68db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2639352
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845956}